+++ src/contents/ui/config/General.qml
@@ -13,8 +13,6 @@ import QtQuick.Controls as QtControls
 import QtQuick.Layouts
 import org.kde.kirigami as Kirigami
 import org.kde.kquickcontrols as KQControls
-import org.kde.plasma.components as PlasmaComponents
-import org.kde.plasma.core as PlasmaCore
 import "../../js/layouts.js" as Layouts
 import "../../js/utils.js" as Utils
 import "../lib"
@@ -53,12 +51,12 @@ Kirigami.FormLayout {
 	}
 
 	RowLayout {
-		PlasmaComponents.CheckBox {
+		QtControls.CheckBox {
 			id: useUserLayout
 			text: i18n("Use user layout")
 		}
 
-		PlasmaComponents.ComboBox {
+		QtControls.ComboBox {
 			id: activeLayoutSlotSelector
 			enabled: cfg_useUserLayout
 			model: [i18n("Slot 1"), i18n("Slot 2"), i18n("Slot 3")]
@@ -78,7 +76,7 @@ Kirigami.FormLayout {
 			value: cfg_flipInterval
 			onValueChanged: cfg_flipInterval = value
 		}
-		PlasmaComponents.Label {
+		QtControls.Label {
 			text: i18n('ms')
 		}
 	}
@@ -94,12 +92,12 @@ Kirigami.FormLayout {
 			value: cfg_randomInterval
 			onValueChanged: cfg_randomInterval = value
 		}
-		PlasmaComponents.Label {
+		QtControls.Label {
 			text: i18n('ms')
 		}
 	}
 
-	PlasmaComponents.CheckBox {
+	QtControls.CheckBox {
 		id: useCustomFont
 		text: i18n("Use custom font")
 	}
@@ -107,10 +105,10 @@ Kirigami.FormLayout {
 		enabled: cfg_useCustomFont
 
 		ColumnLayout {
-			PlasmaComponents.Label {
+			QtControls.Label {
 				text: i18n('Font: %1', cfg_customFont.family)
 			}
-			PlasmaComponents.Label {
+			QtControls.Label {
 				text: i18n('Size: %1', cfg_customFont.pointSize)
 			}
 		}
@@ -147,11 +145,11 @@ Kirigami.FormLayout {
 			enabled: cfg_clockTimezoneOffsetEnabled
 		}
 	}
-	PlasmaComponents.CheckBox {
+	QtControls.CheckBox {
 		id: widgetContainerFillWidth
 		text: i18n("Container fill width")
 	}
-	PlasmaComponents.CheckBox {
+	QtControls.CheckBox {
 		id: widgetContainerFillHeight
 		text: i18n("Container fill height")
 	}
@@ -172,7 +170,7 @@ Kirigami.FormLayout {
 			onValueChanged: cfg_widgetMinWidth = value
 		}
 
-		PlasmaComponents.Label {
+		QtControls.Label {
 			text: i18n('px')
 		}
 	}
@@ -193,7 +191,7 @@ Kirigami.FormLayout {
 			onValueChanged: cfg_widgetMaxWidth = value
 		}
 
-		PlasmaComponents.Label {
+		QtControls.Label {
 			text: i18n('px')
 		}
 	}